requirements.txt
Django==1.10.4
django-extensions==1.7.5
djangorestframework==3.5.3
djoser==0.5.1
drfdocs==0.0.11
msgpack-python==0.4.8
six==1.10.0
